chromium: Create packageconfig to control libc++ to use
Issue: LIN1021-2988

This helps to use either system provided libc++ or bundled libc++ with
chromium. Defaults in chromium is to use bundled version, however in OE
we have used the system provided so far, therefore the default are not
changed here, just an option offered to enable custom libc++ if someone
desires so. Should we default to use the custom copy by default, perhaps
yes but thats a separate discussion

(master rev: 3840a50 )

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>